Updating Templates, Tests, and Scripts for Asciidoctor 2.0
This commit compiles to the changes to our templates (.erb files),
tests, and scripts to work with the changes brought on by Asciidctor 2.0.

* The block_math template had to be renamed to block_stem (no code changes)
* Asciidoctor introduced a code change to strip inline latexmath passthroughs
of "$" delimiters. This interfered with our custom math toolchain so I reworked
the inline_quoted template to check for delimiters and add them if they were
not present. This fixed the issue introduced by Asciidoctor stripping the "$",
but as an added bonus latexmath passthroughs will now work even if no delimiters
are present in the source.
* Separated out the logic for asciimath passthroughs. They're rarely used but they
will also work without delimiters now.
* I added several tests to check the new inline math logic
* Asciidoctor changed the name of the render method to convert, so I changed it
in our code.
* I set compat-mode to true in spec_helper and convert_book. This ensures old
formatting still works. This is import to prevent breaking projects written
using the older version of Asciidoctor.