Updating syntax and dependencies to match Atlas Workers Ruby version
While I was testing this locally, I noticed that there was still Ruby 2.x
era syntax in the scripts and spec files. I went ahead and updated those
and also set the Ruby version in the Gemfile to match what's on the Workers.

EDIT: Updating Ruby version in GitHub CI