Update inline latexmath: handling for latex2mathml in atlas
Change needed, since the \( foo \) syntax confuses the latex2mathml
converter, and frankly, since we're outputting inside a data-type="tex"
span, I think it makes sense to handle that here. If we instead want to
handle that inside latexmath, just say so. I'm pretty agnostic at the
end of the day.